requirements.txt
fastapi
uvicorn[standard]
python-multipart
twilio
openai
requests
pydantic
aiofiles
python-dotenv
openpyxl
gTTS
websockets
python-dotenv
